Rename mode folder and service
